requirements.txt
streamlit==1.31.0
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.4.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.26.3
